requirements.txt
aniso8601==4.1.0
astroid==2.1.0
attrs==19.3.0
autopep8==1.4.4
Click==7.0
cloudpickle==0.7.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
dlib==19.16.0
face-recognition==1.2.3
face-recognition-models==0.3.0
Flask==1.0.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.7
importlib-metadata==1.3.0
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jsonschema==3.2.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
more-itertools==8.0.2
networkx==2.2
numpy==1.15.0
opencv-python==3.4.2.17
Pillow==5.4.1
pycodestyle==2.5.0
pylint==2.2.2
pyparsing==2.3.1
pyrsistent==0.15.6
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
toolz==0.9.0
typed-ast==1.3.0
Werkzeug==0.14.1
wrapt==1.11.1
zipp==0.6.0
